+++ tests/dense.py
@@ -24,9 +24,8 @@
 import unittest
 import torch
 
+import MinkowskiEngine as ME
 from MinkowskiEngine import SparseTensor, MinkowskiConvolution
-
-from tests.common import data_loader
 
 
 class TestDense(unittest.TestCase):
@@ -34,10 +33,25 @@ class TestDense(unittest.TestCase):
     def test(self):
         print(f"{self.__class__.__name__}: test_dense")
         in_channels, out_channels, D = 2, 3, 2
-        coords, feats, labels = data_loader(in_channels)
-        feats = feats.double()
-        feats.requires_grad_()
+        coords1 = torch.IntTensor([[0, 0], [0, 1], [1, 1]])
+        feats1 = torch.DoubleTensor([[1, 2], [3, 4], [5, 6]])
+
+        coords2 = torch.IntTensor([[1, 1], [1, 2], [2, 1]])
+        feats2 = torch.DoubleTensor([[7, 8], [9, 10], [11, 12]])
+        coords, feats = ME.utils.sparse_collate([coords1, coords2], [feats1, feats2])
         input = SparseTensor(feats, coords=coords)
+        input.requires_grad_()
+        dinput, min_coord, tensor_stride = input.dense()
+        self.assertTrue(dinput[0, 0, 0, 1] == 3)
+        self.assertTrue(dinput[0, 1, 0, 1] == 4)
+        self.assertTrue(dinput[0, 0, 1, 1] == 5)
+        self.assertTrue(dinput[0, 1, 1, 1] == 6)
+
+        self.assertTrue(dinput[1, 0, 1, 1] == 7)
+        self.assertTrue(dinput[1, 1, 1, 1] == 8)
+        self.assertTrue(dinput[1, 0, 2, 1] == 11)
+        self.assertTrue(dinput[1, 1, 2, 1] == 12)
+
         # Initialize context
         conv = MinkowskiConvolution(
             in_channels,